+++ pytsite/events/__init__.py
@@ -1,51 +1,71 @@
 """Event Subsystem.
 """
+import re as _re
 from pytsite import reg as _reg, logger as _logger
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-_listeners = {}
+_listeners = {}  # Listeners for exact event names
+_wc_listeners = []  # Listeners for wildcard event names
 _dbg = _reg.get('events.debug', False)
 
 
-def listen(event: str, listener: callable, call_once: bool = False, priority: int = 0):
+def listen(event_name: str, listener: callable, priority: int = 0):
     """Add an event listener.
     """
-    global _listeners
-    if event not in _listeners:
-        _listeners[event] = []
+    global _listeners, _wc_listeners
 
-    _listeners[event].append((listener, call_once, priority))
+    # Exact event name
+    if not event_name.endswith('*'):
+        if event_name not in _listeners:
+            _listeners[event_name] = []
+        _listeners[event_name].append((listener, priority))
+        _listeners[event_name] = sorted(_listeners[event_name], key=lambda x: x[1])  # Sort by priority
 
-    # Sort listeners by priority
-    _listeners[event] = sorted(_listeners[event], key=lambda x: x[2])
+    # Wildcard event name
+    else:
+        re = _re.compile(event_name.replace('.', '\.').replace('*', '.*'))
+        _wc_listeners.append((listener, priority, re))
+        _wc_listeners = sorted(_wc_listeners, key=lambda x: x[1])  # Sort by priority
 
     if _dbg:
-        _logger.debug("Listener attached to event '{}': {}.{}()".format(event, listener.__module__, listener.__name__))
+        _logger.debug('Listener {}.{}() attached to event {}'.format(event_name, listener.__module__, listener.__name__))
 
 
-def fire(event: str, stop_after: int = None, **kwargs) -> list:
+def fire(event_name: str, stop_after: int = None, **kwargs) -> list:
     """Fires an event to listeners.
     """
+    count = 0
+    stop_processing = False
     r = []
 
-    if event not in _listeners:
-        return r
+    # Call listeners for exact event name
+    if event_name in _listeners:
+        for handler, priority in _listeners[event_name]:
+            # Call handler
+            r.append(handler(**kwargs))
 
-    count = 0
-    for handler, call_once, priority in _listeners[event]:
-        # Call handler
-        r.append(handler(**kwargs))
+            # Count called handler and stop if it is necessary
+            count += 1
+            if stop_after and count >= stop_after:
+                stop_processing = True
+                break
 
-        # Count called handler and stop if it is necessary
-        count += 1
-        if stop_after and count >= stop_after:
-            break
+    # Call listeners for exact event name
+    if not stop_processing:
+        for handler, priority, re in _wc_listeners:
+            if not re.match(event_name):
+                continue
 
-    # Remove handlers which should be called once
-    _listeners[event] = [item for item in _listeners[event] if not item[1]]
+            # Call handler
+            r.append(handler(**kwargs))
+
+            # Count called handler and stop if it is necessary
+            count += 1
+            if stop_after and count >= stop_after:
+                break
 
     return r