+++ pytsite/http_api/_api.py
@@ -43,10 +43,10 @@ def call(name: str, inp: dict = None, **args) -> tuple:
 def on_pre_request(handler, priority: int = 0):
     """Register handler which will be called before handling every request to HTTP API.
     """
-    _events.listen('pytsite.http_api.pre_request', handler, priority=priority)
+    _events.listen('pytsite.http_api.pre_request', handler, priority)
 
 
 def on_request(handler, priority: int = 0):
     """Register handler which will be called on every request to HTTP API.
     """
-    _events.listen('pytsite.http_api.request', handler, priority=priority)
+    _events.listen('pytsite.http_api.request', handler, priority)